% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fonts_selector}
\alias{fonts_selector}
\title{Font selector, internal helper function}
\usage{
fonts_selector(...)
}
\arguments{
\item{...}{list of fonts requested}
}
\description{
Will return the first available font
}
