% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{pal}
\alias{pal}
\title{Load a palette in hex definition}
\usage{
pal(palette_name, n = NULL, evenly_spaced = TRUE)
}
\arguments{
\item{palette_name}{the palette's name}

\item{n}{the number of colors to return, will default to the actual size of the palette, will recycle when necessary}

\item{evenly_spaced}{if \code{TRUE}, when the pallete has e.g. 5 colors but only 3 are requested, color 1, 3, and 5 will be returned.}
}
\value{
a character vector of hex colors
}
\description{
See the available palettes using \code{list_palettes}.
\code{pal} will also search for the palette name in \code{RColorBrewer} and return it if found.
}
