% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_economist}
\alias{theme_economist}
\title{Theme 'economist'}
\usage{
theme_economist(grid_lines = "vertical")
}
\arguments{
\item{grid_lines}{the direction of the grid lines, either \code{vertical} or \code{horizontal}}
}
\value{
theme
}
\description{
A theme inspired by https://www.ft.com/
}
