% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_five38}
\alias{theme_five38}
\title{Theme 'five38'}
\usage{
theme_five38(grid_lines = "vertical")
}
\arguments{
\item{grid_lines}{the direction of the grid lines, either \code{vertical} or \code{horizontal}}
}
\value{
theme
}
\description{
A theme inspired by http://fivethirtyeight.com/
}
