% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_monokai_full}
\alias{theme_monokai_full}
\title{Theme 'monokai_full'}
\usage{
theme_monokai_full()
}
\value{
theme
}
\description{
A theme that was strongly based on the base16 monokai scheme http://chriskempson.github.io/base16/
}
