% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_scientific}
\alias{theme_scientific}
\title{Theme 'scientific'}
\usage{
theme_scientific()
}
\value{
theme
}
\description{
A theme that was strongly based on http://minimaxir.com/2015/02/ggplot-tutorial/
}
