% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{watermark_img}
\alias{watermark_img}
\title{Add a watermark image}
\usage{
watermark_img(filename, x, y, width = 32, location = "br", alpha = 1, ...)
}
\arguments{
\item{filename}{the filename of the}

\item{x}{the x location of the watermark (0-1 scale)}

\item{y}{the y location of the watermark (0-1 scale)}

\item{width}{the width of the watermark image (in 'points')}

\item{location}{one of \code{tr} (top right), \code{tl} (top left), \code{bl} (bottom left), \code{br} (bottom right), \code{center} (view center)}

\item{alpha}{transparency value}

\item{...}{arguments passed to \code{grid::gpar()}}
}
\value{
a ggplot object
}
\description{
Add a watermark image
}
