% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{watermark_txt}
\alias{watermark_txt}
\title{Add a watermark text}
\usage{
watermark_txt(txt, x, y, location = "br", rot = 0, ...)
}
\arguments{
\item{txt}{the text to display}

\item{x}{the x location of the watermark (0-1 scale)}

\item{y}{the y location of the watermark (0-1 scale)}

\item{location}{one of \code{tr} (top right), \code{tl} (top left), \code{bl} (bottom left), \code{br} (bottom right), \code{center} (view center)}

\item{rot}{the rotation of the text}

\item{...}{arguments passed to \code{grid::gpar()} (e.g. \code{alpha})}
}
\value{
a ggplot object
}
\description{
Add a watermark text
}
