% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_info}
\alias{palette_info}
\title{Show details of specific palette}
\usage{
palette_info(palette_name)
}
\arguments{
\item{palette_name}{the name of the palette}
}
\description{
Show details of specific palette
}
\examples{
palette_info('flatpaleta')
}
